+++ src/main/java/by/stolybko/Main.java
@@ -1,19 +1,17 @@
 package by.stolybko;
 
+import by.stolybko.api.DeSerializer;
 import by.stolybko.entity.Order;
-import by.stolybko.entity.Product;
-import by.stolybko.parser.Lexeme;
-import by.stolybko.parser.LexemeBuffer;
-
-import java.lang.reflect.Constructor;
+import by.stolybko.api.parser.Lexeme;
+import by.stolybko.api.parser.LexemeBuffer;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
 
-import static by.stolybko.parser.Lexeme.lexAnalyze;
+
+import static by.stolybko.api.parser.ParserJson.lexAnalyze;
 
 public class Main {
 
@@ -44,86 +42,14 @@ static String json = """
 //
 //        System.out.println(isWrapper(field.getType()));
 //        System.out.println(Collection.class.isAssignableFrom(field.getType()));
-        List<Lexeme> lexems = lexAnalyze(json);
-        Order order = (Order) deSerializer(Order.class, lexems);
+
+        DeSerializer deSerializer = new DeSerializer();
+        Order order = deSerializer.deSerializerJson(Order.class, json);
         System.out.println(order);
 
 
     }
-    public static Object deSerializer(Class clazz, List<Lexeme> lexemes) throws NoSuchMethodException, InvocationTargetException, InstantiationException, IllegalAccessException, NoSuchFieldException {
 
-        Object obj = clazz.getConstructor().newInstance();
-        //String[] map = json.split(",");
-
-       /* for (Lexeme l : lexemes) {
-            //String[] kv = s.split(":");
-
-            Field field = clazz.getDeclaredField(kv[0].trim());
-            field.setAccessible(true);
-            Object value = getValue(field, kv[1].trim() );
-
-            field.set(obj, value);
-
-        }*/
-        LexemeBuffer lexemeBuffer = new LexemeBuffer(lexemes);
-        Lexeme lexemeKey = lexemeBuffer.nextKey();
-        do {
-            Field field = clazz.getDeclaredField(lexemeKey.getValue());
-            field.setAccessible(true);
-            //field.getType().isAssignableFrom()
-
-
-            if(isWrapper(field.getType()) || field.getType().isPrimitive()) {
-                Lexeme lexemeValue = lexemeBuffer.nextValue();
-                Object value = getValue(field, lexemeValue.getValue());
-                field.set(obj, value);
-            }
-//            else if (Collection.class.isAssignableFrom(field.getType())) {
-//
-//            }
-            else {
-                Object value = deSerializer(field.getType(), lexemeBuffer.nextObject());
-                field.set(obj, value);
-
-            }
-
-
-
-
-            lexemeKey = lexemeBuffer.nextKey();
-
-        } while (lexemeKey != null);
-
-        return obj;
-
-    }
-
-
-    public static Object getValue(Field field, String value) throws NoSuchFieldException, InvocationTargetException, NoSuchMethodException, InstantiationException, IllegalAccessException {
-        if (field.getType() == UUID.class) {
-            return UUID.fromString(value);
-        }
-        if (field.getType() == Integer.class) {
-            return Integer.valueOf(value);
-        }
-        if (field.getType() == Double.class) {
-            return Double.valueOf(value);
-        }
-        if (field.getType() == String.class) {
-            return value;
-        }
-        return null;
-    }
-
-    public static boolean isWrapper(Class clazz) {
-        List<Class> wrapper = new ArrayList<>();
-        wrapper.add(UUID.class);
-        wrapper.add(Integer.class);
-        wrapper.add(String.class);
-        wrapper.add(Double.class);
-
-        return wrapper.contains(clazz);
-    }
 
 
 }
