+++ src/main/java/by/stolybko/api/parser/Lexeme.java
@@ -1,0 +1,14 @@
+package by.stolybko.api.parser;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+
+@Data
+@AllArgsConstructor
+public class Lexeme {
+
+    LexemeType type;
+    String value;
+
+}
