+++ src/main/java/by/stolybko/api/parser/ParserJson.java
@@ -1,38 +1,16 @@
-package by.stolybko.parser;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
+package by.stolybko.api.parser;
 
 import java.util.ArrayList;
 import java.util.List;
 
-@Data
-@AllArgsConstructor
-public class Lexeme {
-    LexemeType type;
+public class ParserJson {
+    public static List<Lexeme> lexAnalyze(String json) {
 
-    //String key;
-    String value;
-
- /*   START_OBJECT, END_OBJECT,
-
-    START_ARRAY, END_ARRAY,
-
-    KEY_VALUE_SEPARATOR
-    PROPERTY_SEPARATOR
-
-    IS_OBJECT, IS_ARRAY,
-    IS_STRING, IS_BOOLEAN, IS_NUMBER,
-
-    PROPERTY;*/
-
-
-    public static List<Lexeme> lexAnalyze(String expText) {
         ArrayList<Lexeme> lexemes = new ArrayList<>();
         int pos = 0;
         boolean isValue = false;
-        while (pos< expText.length()) {
-            char c = expText.charAt(pos);
+        while (pos< json.length()) {
+            char c = json.charAt(pos);
             switch (c) {
                 case '{':
                     lexemes.add(new Lexeme(LexemeType.START_OBJECT, String.valueOf(c)));
@@ -61,7 +39,7 @@ public class Lexeme {
                     continue;
                 case ',':
                     lexemes.add(new Lexeme(LexemeType.PROPERTY_SEPARATOR, String.valueOf(c)));
-                    isValue = !isValue;
+                    isValue = false;
                     pos++;
                     continue;
                 case 32:
@@ -74,37 +52,19 @@ public class Lexeme {
                     do {
                         sb.append(c);
                         pos++;
-                        if (pos >= expText.length()) {
+                        if (pos >= json.length()) {
                             break;
                         }
-                        c = expText.charAt(pos);
+                        c = json.charAt(pos);
                     } while (c != 32 && c != 34 && c != '[' && c != ']' && c != '{' && c != '}' && c != ',' && c != ':');
                     if(isValue) {
                         lexemes.add(new Lexeme(LexemeType.VALUE, sb.toString()));
                     }
                     else lexemes.add(new Lexeme(LexemeType.KEY, sb.toString()));
 
-                /*    if (c <= '9' && c >= '0') {
-                        StringBuilder sb = new StringBuilder();
-                        do {
-                            sb.append(c);
-                            pos++;
-                            if (pos >= expText.length()) {
-                                break;
-                            }
-                            c = expText.charAt(pos);
-                        } while (c <= '9' && c >= '0');
-                        lexemes.add(new Lexeme(LexemeType.NUMBER, sb.toString()));
-                    } else {
-                        if (c != ' ') {
-                            throw new RuntimeException("Unexpected character: " + c);
-                        }
-                        pos++;
-                    }*/
             }
         }
         lexemes.add(new Lexeme(LexemeType.EOF, ""));
         return lexemes;
     }
-
 }
